Create scripts dir
