+++ src/scripts/index.ts
@@ -1,0 +1,2 @@
+export { setupDemo } from './setupDemo'
+export { createCoreInstance, Environment } from './createCoreInstance'
